refactor EconomicResource app into separate modules
no code changes except for increasing field visibility
