+++ Cargo.toml
@@ -4,7 +4,18 @@ members= [
   "example/knowledge-system-extensions/zomes/beef_resource_specification/code",
   "lib/hdk_graph_helpers",
   "lib/vf_core",
-  "lib/vf_observation",
+  "lib/rea_economic_event/defs",
+  "lib/rea_economic_event/lib",
+  "lib/rea_economic_event/structs_internal",
+  "lib/rea_economic_event/structs_rpc",
+  "lib/rea_economic_resource/defs",
+  "lib/rea_economic_resource/lib",
+  "lib/rea_economic_resource/structs_internal",
+  "lib/rea_economic_resource/structs_rpc",
+  "lib/rea_process/defs",
+  "lib/rea_process/lib",
+  "lib/rea_process/structs_internal",
+  "lib/rea_process/structs_rpc",
   "lib/vf_planning",
   "lib/vf_knowledge",
   "lib/vf_specification",
