+++ happs/observation/zomes/economic_resource/code/.hcbuild
@@ -10,5 +10,5 @@
       ]
     }
   ],
-  "artifact": "/tmp/holochain/target/wasm32-unknown-unknown/release/rea_observation_resource.wasm"
+  "artifact": "/tmp/holochain/target/wasm32-unknown-unknown/release/hc_zome_rea_economic_resource.wasm"
 }
