+++ happs/observation/zomes/economic_resource/code/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rea_observation_resource"
+name = "hc_zome_rea_economic_resource"
 version = "0.1.0"
 authors = ["pospi <pospi@spadgos.com>"]
 edition = "2018"
@@ -14,9 +14,13 @@ hdk_proc_macros = "=0.0.40-alpha1"
 holochain_wasm_utils = "=0.0.40-alpha1"
 holochain_json_derive = "=0.0.17"
 
-hdk_graph_helpers = { path = "../../../../../lib/hdk_graph_helpers" }
-vf_observation = { path = "../../../../../lib/vf_observation" }
 vf_specification = { path = "../../../../../lib/vf_specification" }
+
+hc_zome_rea_economic_resource_defs = { path = "../../../../../lib/rea_economic_resource/defs" }
+hc_zome_rea_economic_resource_lib = { path = "../../../../../lib/rea_economic_resource/lib" }
+hc_zome_rea_economic_resource_structs_rpc = { path = "../../../../../lib/rea_economic_resource/structs_rpc" }
+hc_zome_rea_economic_resource_structs_internal = { path = "../../../../../lib/rea_economic_resource/structs_internal" }
+hc_zome_rea_economic_event_structs_rpc = { path = "../../../../../lib/rea_economic_event/structs_rpc" }
 
 [lib]
 path = "src/lib.rs"
