+++ happs/observation/zomes/economic_resource/code/src/lib.rs
@@ -1,48 +1,18 @@
 #![feature(proc_macro_hygiene)]
-// :TODO: documentation
-
-extern crate hdk;
-extern crate serde;
-#[macro_use]
-extern crate serde_derive;
-extern crate serde_json;
-extern crate hdk_graph_helpers;
-extern crate vf_observation;
-extern crate vf_specification;
-
-mod economic_resource_requests;
 
 use hdk::prelude::*;
 use hdk_proc_macros::zome;
 
-use vf_observation::type_aliases::{
-    ResourceAddress,
-};
-use vf_observation::economic_resource::{
-    Entry,
-    UpdateRequest,
-    ResponseData,
-};
-use economic_resource_requests::{
-    QueryParams,
-    receive_get_economic_resource,
-    receive_update_economic_resource,
-    receive_query_economic_resources,
-};
-use vf_observation::identifiers::{
-    RESOURCE_BASE_ENTRY_TYPE,
-    RESOURCE_INITIAL_ENTRY_LINK_TYPE,
-    RESOURCE_ENTRY_TYPE,
-    RESOURCE_CONTAINS_LINK_TYPE,
-    RESOURCE_CONTAINED_IN_LINK_TYPE,
-    RESOURCE_AFFECTED_BY_EVENT_LINK_TYPE,
-    EVENT_BASE_ENTRY_TYPE,
-    RESOURCE_CONFORMS_TO_LINK_TYPE,
-};
 use vf_specification::identifiers::{
     ECONOMIC_RESOURCE_SPECIFICATION_BASE_ENTRY_TYPE,
     RESOURCE_SPECIFICATION_CONFORMING_RESOURCE_LINK_TYPE,
 };
+use hc_zome_rea_economic_resource_structs_internal::identifiers::RESOURCE_BASE_ENTRY_TYPE;
+
+use hc_zome_rea_economic_resource_defs::{ entry_def, base_entry_def };
+use hc_zome_rea_economic_resource_lib::*;
+use hc_zome_rea_economic_resource_structs_rpc::*;
+use hc_zome_rea_economic_event_structs_rpc::ResourceResponseData as ResponseData;
 
 #[zome]
 mod rea_economic_resource_zome {
@@ -59,102 +29,12 @@ mod rea_economic_resource_zome {
 
     #[entry_def]
     fn resource_entry_def() -> ValidatingEntryType {
-        entry!(
-            name: RESOURCE_ENTRY_TYPE,
-            description: "A resource which is useful to people or the ecosystem.",
-            sharing: Sharing::Public,
-            validation_package: || {
-                hdk::ValidationPackageDefinition::Entry
-            },
-            validation: |validation_data: hdk::EntryValidationData<Entry>| {
-                // CREATE
-                if let EntryValidationData::Create{ entry, validation_data: _ } = validation_data {
-                    let record: Entry = entry;
-                    return record.validate();
-                }
-
-                // UPDATE
-                if let EntryValidationData::Modify{ new_entry, old_entry: _, old_entry_header: _, validation_data: _ } = validation_data {
-                    let record: Entry = new_entry;
-                    return record.validate();
-                }
-
-                // DELETE
-                // if let EntryValidationData::Delete{ old_entry, old_entry_header: _, validation_data: _ } = validation_data {
-
-                // }
-
-                Ok(())
-            }
-        )
-
+        entry_def()
     }
 
     #[entry_def]
     fn resource_base_entry_def() -> ValidatingEntryType {
-        entry!(
-            name: RESOURCE_BASE_ENTRY_TYPE,
-            description: "Base anchor for initial resource addresses to provide lookup functionality",
-            sharing: Sharing::Public,
-            validation_package: || {
-                hdk::ValidationPackageDefinition::Entry
-            },
-            validation: |_validation_data: hdk::EntryValidationData<Address>| {
-                Ok(())
-            },
-            links: [
-                to!(
-                    RESOURCE_ENTRY_TYPE,
-                    link_type: RESOURCE_INITIAL_ENTRY_LINK_TYPE,
-                    validation_package: || {
-                        hdk::ValidationPackageDefinition::Entry
-                    },
-                    validation: | _validation_data: hdk::LinkValidationData| {
-                        Ok(())
-                    }
-                ),
-                to!(
-                    RESOURCE_BASE_ENTRY_TYPE,
-                    link_type: RESOURCE_CONTAINS_LINK_TYPE,
-                    validation_package: || {
-                        hdk::ValidationPackageDefinition::Entry
-                    },
-                    validation: | _validation_data: hdk::LinkValidationData| {
-                        Ok(())
-                    }
-                ),
-                to!(
-                    RESOURCE_BASE_ENTRY_TYPE,
-                    link_type: RESOURCE_CONTAINED_IN_LINK_TYPE,
-                    validation_package: || {
-                        hdk::ValidationPackageDefinition::Entry
-                    },
-                    validation: | _validation_data: hdk::LinkValidationData| {
-                        Ok(())
-                    }
-                ),
-                to!(
-                    EVENT_BASE_ENTRY_TYPE,
-                    link_type: RESOURCE_AFFECTED_BY_EVENT_LINK_TYPE,
-                    validation_package: || {
-                        hdk::ValidationPackageDefinition::Entry
-                    },
-                    validation: | _validation_data: hdk::LinkValidationData| {
-                        Ok(())
-                    }
-                ),
-                to!(
-                    ECONOMIC_RESOURCE_SPECIFICATION_BASE_ENTRY_TYPE,
-                    link_type: RESOURCE_CONFORMS_TO_LINK_TYPE,
-                    validation_package: || {
-                        hdk::ValidationPackageDefinition::Entry
-                    },
-                    validation: | _validation_data: hdk::LinkValidationData| {
-                        Ok(())
-                    }
-                )
-            ]
-        )
+        base_entry_def()
     }
 
     // :TODO: move to separate zome
