+++ lib/rea_economic_resource/defs/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "hc_zome_rea_economic_resource_defs"
+version = "0.1.0"
+authors = ["pospi <pospi@spadgos.com>"]
+edition = "2018"
+
+[dependencies]
+# :DUPE: hdk-rust-revid
+hdk = "=0.0.40-alpha1"
+
+vf_specification = { path = "../../vf_specification" }
+hc_zome_rea_economic_resource_structs_internal = { path = "../structs_internal" }
+hc_zome_rea_economic_event_structs_internal = { path = "../../rea_economic_event/structs_internal" }
+
+[lib]
+crate-type = ["lib"]
