+++ lib/rea_economic_resource/lib/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "hc_zome_rea_economic_resource_lib"
+version = "0.1.0"
+authors = ["pospi <pospi@spadgos.com>"]
+edition = "2018"
+
+[dependencies]
+# :DUPE: hdk-rust-revid
+hdk = "=0.0.40-alpha1"
+
+hdk_graph_helpers = { path = "../../hdk_graph_helpers" }
+vf_core = { path = "../../vf_core" }
+vf_specification = { path = "../../vf_specification" }
+hc_zome_rea_economic_resource_structs_internal = { path = "../structs_internal" }
+hc_zome_rea_economic_resource_structs_rpc = { path = "../structs_rpc" }
+hc_zome_rea_economic_event_structs_internal = { path = "../../rea_economic_event/structs_internal" }
+hc_zome_rea_economic_event_structs_rpc = { path = "../../rea_economic_event/structs_rpc" }
+
+[lib]
+crate-type = ["lib"]
